Comic pages were being stored in the wrong order.
I've changed the data structure that stores the pages and their source files from a map to two slices. Maps in Go are not guaranteed to be ordered so the order the pages were sent in was lost when it came to storing them in the db (the order doesn't matter at the filestore). This wasn't apparent in testing because it seems go test keeps the order of maps whereas the go runtime purposely randomises the order to force people to deal with maps being unordered.