Changed the representation of links in the HTTP API.
The key (or rel) for a link is now contained within the link object itself rather than as a key of the parent object. The difficulty with using the rel as a key is that a dictionary can't represent the collection when there are multiple links with the same rel because dictionaries expect there to be a single value for a unique key.